+++ CHANGELOG.md
@@ -54,6 +54,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed error when duplicate canvases where imported into single manifest
 - Moved media helpers to shared modules
 - Fixed unhandled exception in top contributor page block
+- Fixed bug with checking if canvas inside of manifest
 
 ## [1.1.3](https://github.com/digirati-co-uk/madoc-platform/compare/v1.1.2...v1.1.3) - 2019-06-03